refactor(feed): extract video_feed_page widget classes (#4510)
Move the home-feed empty/error state widgets out of video_feed_page.dart
into a colocated video_feed_page/ folder, mirroring the explore_screen
split. The page file drops from 617 to 427 lines, clearing the <500-line
acceptance bar from #4510.

#4682's mediakit->native migration already deleted the pooled-player
widget classes the issue originally named (1,442->617) but left the file
just over the line target and never linked the issue. This finishes it:

- feed_empty_widget.dart: FeedEmptyWidget + Following empty state + mark
- feed_error_widget.dart: FeedErrorWidget (was private _FeedErrorWidget)

The empty-state icon moves from raw Icons.video_library_outlined to
DivineIcon(filmSlate): the raw-Icons ratchet forbids a raw Material icon
in a new file (ui_theming.md mandates DivineIcon), so the widget can't be
relocated verbatim. Baseline shrinks by one (video_feed_page.dart row).

Existing feed-page tests pass unchanged.
