+++ mobile/test/screens/feed/video_feed_page_test.dart
@@ -23,6 +23,7 @@ import 'package:openvine/providers/shell_obscured_provider.dart';
 import 'package:openvine/router/router.dart';
 import 'package:openvine/screens/explore/explore_screen.dart';
 import 'package:openvine/screens/feed/video_feed_page.dart';
+import 'package:openvine/screens/feed/video_feed_page/feed_empty_widget.dart';
 import 'package:openvine/services/view_event_publisher.dart';
 import 'package:openvine/widgets/video_feed_item/feed_videos.dart';
